ci: refactor tests
* move `integration` dir into `tests`
* create subdirectories in `tests`
* split `integration_tests` into
  `c_integration_tests` and `rust_integration_tests`
