+++ Cargo.toml
@@ -79,5 +79,5 @@ walkdir = "2"
 wat = { version = "1.0" }
 
 [workspace]
-members = [ "integration/sev_attestation", "integration/simple", "integration/wasm" ]
+members = [ "tests/sev_attestation", "tests/rust-exec", "tests/wasm/rust-tests" ]
 exclude = [ "internal/shim-sev", "internal/shim-sgx", "internal/wasmldr" ]
