+++ tests/c-tests/close.c
