+++ tests/c-tests/exit_zero.c
