+++ tests/c-tests/sgx_get_att_quote.c
