+++ tests/c-tests/uname.c
