+++ tests/c-tests/write_emsgsize.c
