+++ tests/c-tests/write_stderr.c
