+++ tests/c-tests/write_stdout.c
