+++ tests/rust-exec/src/cpuid.rs
