+++ tests/rust-exec/src/echo.rs
