+++ tests/rust-exec/src/memory_stress_test.rs
