+++ tests/rust-exec/src/memspike.rs
