+++ tests/rust-exec/src/unix_echo.rs
