+++ tests/wasm/rust-tests/.cargo/config
@@ -2,4 +2,4 @@
 target = "wasm32-wasi"
 
 [target.wasm32-wasi]
-runner = "../../helper/test-enarx-wasm.sh"
+runner = "../../../helper/test-enarx-wasm.sh"
