+++ tests/wasm/rust-tests/rust-toolchain.toml
