+++ tests/wasm/rust-tests/src/echo.rs
