+++ tests/wasm/rust-tests/src/memspike.rs
