+++ tests/wasm/rust-tests/src/zerooneone.rs
